Check if the file exists before checking its size

+++ cowrie/insults/insults.py
@@ -179,20 +179,21 @@ class LoggingServerProtocol(insults.ServerProtocol):
 
         if self.redirlogOpen:
             try:
-                if os.path.getsize(self.redirlogFile) > 0:
-                    with open(self.redirlogFile, 'rb') as f:
-                        shasum = hashlib.sha256(f.read()).hexdigest()
-                        shasumfile = self.downloadPath + "/" + shasum
-                        if os.path.exists(shasumfile):
-                            os.remove(self.redirlogFile)
-                        else:
-                            os.rename(self.redirlogFile, shasumfile)
-                        os.symlink(shasum, self.redirlogFile)
-                    log.msg(eventid='cowrie.session.file_download',
-                            format='Saved redir contents to %(outfile)s',
-                            url='redir',
-                            outfile=shasumfile,
-                            shasum=shasum)
+                if os.path.exists(self.redirlogFile):
+                    if os.path.getsize(self.redirlogFile) > 0:
+                        with open(self.redirlogFile, 'rb') as f:
+                            shasum = hashlib.sha256(f.read()).hexdigest()
+                            shasumfile = self.downloadPath + "/" + shasum
+                            if os.path.exists(shasumfile):
+                                os.remove(self.redirlogFile)
+                            else:
+                                os.rename(self.redirlogFile, shasumfile)
+                            os.symlink(shasum, self.redirlogFile)
+                        log.msg(eventid='cowrie.session.file_download',
+                                format='Saved redir contents to %(outfile)s',
+                                url='redir',
+                                outfile=shasumfile,
+                                shasum=shasum)
             except IOError as e:
                 pass
             finally:
